requirements.txt
annotated-types==0.7.0
anyio==4.13.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
distro==1.9.0
Flask==3.1.3
groq==1.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==12.2.0
pydantic==2.13.0
pydantic_core==2.46.0
python-dotenv==1.2.2
reportlab==4.4.10
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.8

